Delete docs directory
The document in this directory was for computing the online covariance matrix.  I realized that when computing the GRM, the QTL2 software does not use the covariance but instead only the second order term.  I've updated to main function to only compute the second order term, so this document is out-of-date with the code base.